In many cases, grastate.dat zeroed if we encounter an error at the node startup. For example, grastate.dat zeroed on misconfiguration in the my.cnf. As a result, we have to perform a full SST after the node restart, although we could only do IST.
This is because the function ReplicatorSMM::request_state_transfer() always marks the state as "unsafe" fearing that current state will be broken in the process of execution of the SST. Since one general function is responsible for the SST and IST, it happens even when the SST is not required or when the SST is trivial.

We must mark the state as the "unsafe" before SST because the current state may be changed during execution of the SST and it will no longer match the stored seqno (state becomes "unsafe" after the first modification of data during the SST, but unfortunately, we do not have callback or wsrep API to notify about the first data modification).

On the other hand, in cases where the full SST is not required and we want to use IST, we need to save the current state - to prevent unnecessary SST after node restart (if IST will be failed).

Therefore, to correct this error we need to check whether the full state transfer (SST) is required or not, before marking state as unsafe.