- #PXC-594: Assertion `sender->count_last_applied' failed in gcs_group_handle_join_msg(gcs_group_t*, const gcs_recv_msg_t*)
This patch fixes two bugs:

PART I, PXC-594:
----------------

Assertion `sender->count_last_applied' failed in the
gcs_group_handle_join_msg() function.

This error stems from the fact that we have added
a direct transition from JOINED to DONOR state bypassing
SYNCED state, but does not take into account that in this
case the count_last_applied flag is lost. It is necessary
to change it in such a direct transition.

PART II, PXC-608:
-----------------

Assertion `node->desync_count > 0' failed in the
the gcs_node_update_status() function.

This error occurs because although disconnection node
from the cluster involves stop of replication, closing
of all client connections and consequently the cancellation
of all transactions (which are tied to client connections),
but the desynchronization counter associated with the current
node, as well as the last state of the node, which is stored
in the prim_state field of the gcs_group_t structure, are not
cleared even after disconnecting the node from the cluster.