PXC-4316: Nodes "changing identity" can prevents primary groups
https://perconadev.atlassian.net/browse/PXC-4316

Problem:
If the node is partitioned from the cluster, and during this time it
shuts down and starts again and then rejoins the cluster, the other part
of the cluster still waits for partitioned node.

Cause:
Node's uuid is maintained in gvwstate.dat file. If the node shuts down
gracefully, the file is deleted. When the node starts and there is no
gvwstate.dat file, new uuid is assigned to this node.
When such a node rejoins the cluster, it is identified by new uuid.
gmcast layer properly detects that the node changed its identity, but
this information is not propagated to upper protocol layers.
pc layer maintains its own set of known nodes. As the information about
identity changing is not delivered to this layer, it sees the "new" node
identified by new uuid, but also waits for "old" node identified by old
uuid.

Solution:
Propagate information about node's identity change to upper protocol
layers.