PXC-4369: PXC 8.0.36 refresh - Q1 2024 (Disable appending zero-level certification keys for NBO and TOI transactions)
Analysis:

The zero-level keys introduced in commit 9439a735 conflicts with our
implementation of NBO certification resulting in certification failure.

In our implementation, we perform NBO certification even for regular
transactions and this makes even the non-conflicting transations to fail
when there is an active NBO running in the cluster.

Fix:
Disable appending zero-level certification keys for NBO and TOI
transactions.