PXC-4428: Garbd terminates SST script's children after SST is complete
https://perconadev.atlassian.net/browse/PXC-4428

Problem:
garbd terminates recv-script immediately after Donor moves from 'donor'
state after serving SST. As the result, if there are any post-receive
actions in recv-script they are not executed (or may be terminated in
the middle)

Cause:
garbd monitors script's error output and when the pipe is closed,
it considers script as finished (err_log_thd). At the same time garbd
monitors the state of the donor with 1 second frequency (sst_status_thd)
The idea is that if script exits after SST, the exit will be detected
by err_log_thd, and sst_status_thd will be stopped.
On the other hand, if recv-script hangs, and Donor exits 'donor' state,
such situation will be detected by sst_status_thd, and script will be
killed. Killing of script will be detected by err_log_thd.
The problem is, that there is no way to tell sst_status_thd that scrit
received data and there is more time needed. The logic is simple:
if Donor exited 'donor' state and the script is still running - kill the
script.

Solution:
1. After detection of Donor exiting 'donor' state, wait for at most
5 seconds for recv-script to finish. In most cases, this will be
immediate action. If the script doesn't exit in 5 seconds, kill it and
consider SST to be failed.
2. Introduced post-recv-script optional parameter. Any post-download
actions should be done here. The script is executed only if recv-script
finished with success (exit 0, was not terminated)